update v1.0
This is an updated version for the pipeline. However, due to the
unstability of Sapelo platform, I plan to make some change to record all
stderr and stdout files. Those files will record if each group job in
each step is completed correctly. In order to make such changes, I first
need to save currrent version.