update v1.1
I believe the stderr and stdout file system is stable in this version
and is actually useful in debugging, so I'm going to integrate this
system in later version.
For this version, some bugs during retrieve reads from bowtie are fixed.
Now I plan to improve on retrieve reads from DIAMON and test how it will
get compatible with following runs. Back up this version first.